requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-webpack-loader==0.6.0
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.3
pandas==0.22.0
patsy==0.5.0
PyJWT==1.6.1
python-dateutil==2.7.2
pytz==2018.3
requests==2.18.4
scipy==1.1.0
six==1.11.0
statsmodels==0.8.0
uritemplate==3.0.0
urllib3==1.22
